Section gauges leave HumansMetricsService; audience segmentation moves to Users (nobodies-collective/Humans#1091)
- New Humans.Base.Hosting.PolledGaugeService: shared 60s gauge-refresh base
  (StartAsync arming, per-refresh scope, error-swallow)
- Gauges move to the section owning the service they read: Governance
  (asociados, applications_pending, pending_consents,
  consent_deadline_approaching), Auth (role_assignments_active), Teams (teams,
  team_join_requests_pending), GoogleIntegration (google_resources,
  google_sync_outbox_pending), Consent (legal_documents_active). Gauge
  names/tags unchanged — dashboards unaffected
- Web's HumansMetricsService subclasses the base, keeping counters + the
  Users/activity gauges only
- GetAudienceSegmentationAsync + AudienceSegmentation move from the database
  diagnostics service onto IAdminDashboardService/Users' AdminDashboardService;
  diagnostics keeps only migration status + Hangfire locks
