+++ src/Sections/Humans.Auth/Section.cs
@@ -85,5 +85,9 @@ public sealed class Section : ISection
         services.AddScoped<IMagicLinkUrlBuilder, MagicLinkUrlBuilder>();
         services.AddScoped<IMagicLinkRateLimiter, MagicLinkRateLimiter>();
         services.AddScoped<IMagicLinkService, MagicLinkService>();
+
+        // Gauge-refresh loop split out of HumansMetricsService (nobodies-collective/Humans#1091).
+        services.AddSingleton<AuthMetricsService>();
+        services.AddHostedService(sp => sp.GetRequiredService<AuthMetricsService>());
     }
 }
